Issue #61

+++ webapp/animal_util.go
@@ -11,7 +11,7 @@ func RandomAnimal(r JSRand) AnimalBuilder {
 	a.SetName(RandomText(r, 20, false))
 	a.SetSummary(RandomText(r, Ternary(false, 300, 20), false))
 	a.SetDetails(RandomText(r, Ternary(false, 2000, 20), true))
-	a.SetCampaignTarget(int((r.Intn(10) + 2) * 50 * DollarsToCurrency))
+	a.SetCampaignTarget((r.Intn(10) + 2) * 50 * DollarsToCurrency)
 	a.SetCampaignBalance(r.Intn(a.CampaignTarget()))
 
 	{
